--- src/main/java/com/kustacks/kuring/feedback/common/dto/SaveFeedbackV1Request.java
@@ -1,21 +1,0 @@
-package com.kustacks.kuring.feedback.common.dto;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.validation.constraints.NotBlank;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class SaveFeedbackV1Request {
-
-    @NotBlank
-    private String id;
-
-    @NotBlank
-    private String content;
-}
-
