+++ src/main/java/com/kustacks/kuring/user/common/dto/SaveFeedbackResponse.java
@@ -1,4 +1,4 @@
-package com.kustacks.kuring.feedback.common.dto;
+package com.kustacks.kuring.user.common.dto;
 
 import com.kustacks.kuring.common.dto.ResponseDto;
 import lombok.Getter;
