+++ src/main/java/com/kustacks/kuring/user/domain/User.java
@@ -1,7 +1,6 @@
 package com.kustacks.kuring.user.domain;
 
 import com.kustacks.kuring.notice.domain.CategoryName;
-import com.kustacks.kuring.feedback.domain.Feedback;
 import com.kustacks.kuring.notice.domain.DepartmentName;
 import lombok.AccessLevel;
 import lombok.Getter;
