+++ src/test/java/com/kustacks/kuring/acceptance/FeedbackStep.java
@@ -1,9 +1,10 @@
 package com.kustacks.kuring.acceptance;
 
-import com.kustacks.kuring.feedback.common.dto.SaveFeedbackV1Request;
+import com.kustacks.kuring.user.common.dto.SaveFeedbackRequest;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -13,18 +14,19 @@ public class FeedbackStep {
 
     public static void 피드백_요청_응답_확인(ExtractableResponse<Response> response) {
         assertAll(
-                () -> assertThat(response.jsonPath().getBoolean("isSuccess")).isTrue(),
-                () -> assertThat(response.jsonPath().getString("resultMsg")).isEqualTo("성공"),
-                () -> assertThat(response.jsonPath().getInt("resultCode")).isEqualTo(201)
+                () -> assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value()),
+                () -> assertThat(response.jsonPath().getInt("code")).isEqualTo(200),
+                () -> assertThat(response.jsonPath().getString("message")).isEqualTo("피드백 저장에 성공하였습니다")
         );
     }
 
     public static ExtractableResponse<Response> 피드백_요청(String fcmToken, String feedback) {
         return RestAssured
                 .given().log().all()
+                .header("User-Token", fcmToken)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(new SaveFeedbackV1Request(fcmToken, feedback))
-                .when().post("/api/v1/feedback")
+                .body(new SaveFeedbackRequest(feedback))
+                .when().post("/api/v2/users/feedbacks")
                 .then().log().all()
                 .extract();
     }
